Added Common OpenAI Tool that shares between Groq, OpenAI and LocalAI
