+++ src/main/kotlin/robot/provider/models/OpenAI.kt
@@ -17,7 +17,6 @@ import dev.supachain.robot.tool.ToolChoice
 import dev.supachain.robot.tool.ToolConfig
 import dev.supachain.robot.tool.strategies.BackAndForth
 import dev.supachain.robot.tool.strategies.ToolUseStrategy
-import dev.supachain.utilities.Parameter
 import dev.supachain.utilities.toJson
 import io.ktor.http.*
 import kotlinx.serialization.SerialName
@@ -139,7 +138,6 @@ class OpenAI : Provider<OpenAI>(), OpenAIActions, OpenAIModels {
 ██████████████████████████████████████████████        ██  ███████████  █████████████████████████████████████████████████
 ██████████████████████████████████████████████  ████  ██  ████████        ██████████████████████████████████████████████
 */
-private fun List<ToolConfig>.asOpenAITools() = map { OpenAIAPI.Tool(it) }
 
 interface OpenAIAPI {
 
@@ -155,33 +153,6 @@ interface OpenAIAPI {
         constructor(message: Message) : this(message.role(), message.text().value)
     }
 
-    @Serializable
-    data class Tool(val type: String, val function: Function) {
-        @Serializable
-        data class Function(val name: String, val description: String?, val parameters: Parameters) {
-            @Serializable
-            data class Parameters(
-                val type: String,
-                val properties: Map<String, Property>,
-                val required: List<String>
-            ) {
-                constructor(parameters: List<Parameter>) : this(
-                    "object",
-                    parameters.toProperties(),
-                    parameters.filter { it.required }.map { it.name }
-                )
-            }
-
-            constructor(toolConfig: ToolConfig) :
-                    this(
-                        toolConfig.function.name,
-                        toolConfig.function.description.ifBlank { null },
-                        Parameters(toolConfig.function.parameters)
-                    )
-        }
-
-        constructor(toolConfig: ToolConfig) : this("function", Function(toolConfig))
-    }
 
     /*
     ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░      ░░░  ░░░░  ░░░      ░░░        ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
@@ -206,7 +177,7 @@ interface OpenAIAPI {
         val stream: Boolean,
         val temperature: Double,
         @SerialName("top_p") val topP: Double,
-        val tools: List<Tool>,
+        val tools: List<CommonTool.OpenAI>,
         @SerialName("tool_choice") val toolChoice: ToolChoice,
         val user: String?
     ) : CommonChatRequest
